feat: Add generic RequestResult<T> class for enhanced request handling
